requirements.txt
flask
torch
transformers
pillow
torchvision
einops